+++ zcash-swift-wallet-sdk/CHANGELOG.md
@@ -6,6 +6,49 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Unreleased
 
+# 2.4.5 - 2026-03-06
+
+## Fixed
+- Fix for a long-standing note commitment tree corruption error.
+
+## Checkpoints
+
+Mainnet
+
+````
+Sources/ZcashLightClientKit/Resources/checkpoints/mainnet/3252500.json
+...
+Sources/ZcashLightClientKit/Resources/checkpoints/mainnet/3262500.json
+````
+
+Testnet
+
+````
+Sources/ZcashLightClientKit/Resources/checkpoints/testnet/3870000.json
+...
+Sources/ZcashLightClientKit/Resources/checkpoints/testnet/3880000.json
+````
+
+# 2.4.4 - 2026-02-24
+
+## Checkpoints
+
+Mainnet
+
+````
+Sources/ZcashLightClientKit/Resources/checkpoints/mainnet/3220000.json
+...
+Sources/ZcashLightClientKit/Resources/checkpoints/mainnet/3250000.json
+````
+
+Testnet
+
+````
+Sources/ZcashLightClientKit/Resources/checkpoints/testnet/3810000.json
+...
+Sources/ZcashLightClientKit/Resources/checkpoints/testnet/3860000.json
+````
+
 # 2.4.3 - 2026-02-02
 
 ## Fixed
@@ -28,6 +71,7 @@ Sources/ZcashLightClientKit/Resources/checkpoints/testnet/3740000.json
 ...
 Sources/ZcashLightClientKit/Resources/checkpoints/testnet/3800000.json
 ````
+
 # 2.4.2 - 2025-12-16
 
 ## Added